Урок 5. Небольшой рефактор кода.
