+++ app/src/main/java/com/example/theweatherwithnesterenko/view/MainActivity.kt
@@ -1,21 +1,8 @@
 package com.example.theweatherwithnesterenko.view
 
-
-
 import android.os.Bundle
-import android.os.Looper
-import android.util.Log
-import android.view.View
-import android.widget.Button
-import android.widget.LinearLayout
-import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.example.theweatherwithnesterenko.R
-import com.example.theweatherwithnesterenko.lesson3.Lesson3
-import com.example.theweatherwithnesterenko.lesson3.someViewGroup
-import com.example.theweatherwithnesterenko.lesson4.BaseImpl
-import com.example.theweatherwithnesterenko.lesson4.BossDelegate
-import com.example.theweatherwithnesterenko.lesson4.Lesson4
 import com.example.theweatherwithnesterenko.view.weatherlist.WeatherListFragment
 
 
@@ -27,64 +14,5 @@ class MainActivity : AppCompatActivity() {
             supportFragmentManager.beginTransaction()
                 .replace(R.id.container, WeatherListFragment.newInstance()).commit()
         }
-
     }
-
-
-
-
-
-
-
-    //todo после прочтения сжечь!
-
-    // Ниже идёт всякая учебная штуковина, которая не должна влиять на работу приложения.
-    private fun doLesson4Demo() {
-        /*val t = 1
-        val any:Any = t
-        val object1:Objects = t
-        val any1:Any = object1
-        val object2:Objects = any*/
-
-        val button = Button(this)
-        val view1: View = LinearLayout(this)
-        val view2: View = TextView(this)
-        (view2 as TextView).text = ""
-        someViewGroup((view1 as LinearLayout))
-
-
-        val looperNotNullable: Looper = getMainLooper()
-        val looperNullable: Looper? = getMainLooper()
-
-        val lesson3 = Lesson3()
-        val lesson4 = Lesson4()
-        with(lesson4) {
-            this.lesson3 = lesson3
-            some1()//1 способ
-            f = lesson3.f//2 способ
-            some2()
-            speakable = lesson3 //3.1 способ
-            some3()
-            speakable = lesson3.callback //3.2 способ
-            some4()
-            speakable = lesson3.callbackLambda1 //4.1 способ
-            some5()
-            //some6(lesson3.callbackLambda2)
-            some6 { string: String, i: Int ->
-                Log.d("@@@", " Сообщение $string")
-                1.0
-            }
-            was()
-            //main(this@MainActivity)
-        }
-
-        val worker = BaseImpl()
-        BossDelegate(worker, worker).run {
-            manipulate()
-        }
-    }
-
-    private fun Lesson4.was() { // функция-расширение из 4-го урока
-        Log.d("@@@", "Был ${this.pr}")
-    }
-}// END of MainActivity
+}
