+++ app/src/main/java/com/example/theweatherwithnesterenko/view/MainActivityWebView.kt
@@ -1,6 +1,5 @@
 package com.example.theweatherwithnesterenko.view
 
-
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
@@ -21,8 +20,7 @@ class MainActivityWebView : AppCompatActivity() {
         setContentView(binding.root)
         binding.btnSearch.setOnClickListener {
             val urlText = binding.edTextUrl.text.toString()
-            createAndOpenUrlConnection(urlText) // подробная роспись
-//           binding.webView.loadUrl(urlText) // webView сама может всё
+            createAndOpenUrlConnection(urlText)
         }
     }
 
@@ -34,14 +32,10 @@ class MainActivityWebView : AppCompatActivity() {
                 readTimeout = 1000
             }
         Thread {
-            val headers = urlConnection.headerFields // зачем это свойство?
+            val headers = urlConnection.headerFields
             val buffer = BufferedReader(InputStreamReader(urlConnection.inputStream))
             val result = getLinesAsOneBigString(buffer)
-            /*runOnUiThread { // 1 способ
-                    binding.webView.loadData(result, "text/html; utf-8", "utf-8")
-                }*/
-            Handler(Looper.getMainLooper()).post { // 2 способ
-//                binding.webView.settings.javaScriptEnabled = true
+            Handler(Looper.getMainLooper()).post {
                 binding.webView.loadDataWithBaseURL(
                     null,
                     result,
