+++ app/src/main/java/com/example/theweatherwithnesterenko/viewmodel/MainViewModel.kt
@@ -1,12 +1,10 @@
 package com.example.theweatherwithnesterenko.viewmodel
 
-
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.example.theweatherwithnesterenko.repository.Repository
 import com.example.theweatherwithnesterenko.repository.RepositoryImpl
-import java.lang.Thread.sleep
+
 
 class MainViewModel(
     private val liveData: MutableLiveData<AppState> = MutableLiveData(),
@@ -21,7 +19,6 @@ class MainViewModel(
     fun getWeatherRussia() = getWeather(true)
     fun getWeatherWorld() = getWeather(false)
 
-
     private fun getWeather(isRussian: Boolean) {
         Thread {
             liveData.postValue(AppState.Loading)
@@ -34,5 +31,4 @@ class MainViewModel(
                 liveData.postValue(AppState.Error(IllegalAccessException()))
         }.start()
     }
-
 }
